+++ layouts/src/main/java/org/patternfly/layout/flex/Flex.java
@@ -27,6 +27,8 @@ import elemental2.dom.HTMLElement;
 
 import static org.jboss.elemento.Elements.div;
 import static org.patternfly.core.Tuple.tuple;
+import static org.patternfly.style.Breakpoint.default_;
+import static org.patternfly.style.Breakpoints.breakpoints;
 import static org.patternfly.style.Classes.flex;
 import static org.patternfly.style.Classes.item;
 import static org.patternfly.style.Classes.layout;
@@ -66,9 +68,23 @@ public class Flex extends BaseLayout<HTMLElement, Flex> {
 
     // ------------------------------------------------------ builder
 
+    /**
+     * Same as {@code align(breakpoints(default_, align))}
+     */
+    public Flex align(Align align) {
+        return align(breakpoints(default_, align));
+    }
+
     /** Value to use for margin: auto at various breakpoints */
     public Flex align(Breakpoints<Align> align) {
         return css(align.modifiers());
+    }
+
+    /**
+     * Same as {@code alignContent(breakpoints(default_, alignContent))}
+     */
+    public Flex alignContent(AlignContent alignContent) {
+        return alignContent(breakpoints(default_, alignContent));
     }
 
     /** Value to add for align-content property at various breakpoints */
@@ -76,9 +92,23 @@ public class Flex extends BaseLayout<HTMLElement, Flex> {
         return css(alignContent.modifiers());
     }
 
+    /**
+     * Same as {@code alignItems(breakpoints(default_, alignItems))}
+     */
+    public Flex alignItems(AlignItems alignItems) {
+        return alignItems(breakpoints(default_, alignItems));
+    }
+
     /** Value to add for align-items property at various breakpoints */
     public Flex alignItems(Breakpoints<AlignItems> alignItems) {
         return css(alignItems.modifiers());
+    }
+
+    /**
+     * Same as {@code alignSelf(breakpoints(default_, alignSelf))}
+     */
+    public Flex alignSelf(AlignSelf alignSelf) {
+        return alignSelf(breakpoints(default_, alignSelf));
     }
 
     /** Value to add for align-self property at various breakpoints */
@@ -86,9 +116,23 @@ public class Flex extends BaseLayout<HTMLElement, Flex> {
         return css(alignSelf.modifiers());
     }
 
+    /**
+     * Same as {@code columnGap(breakpoints(default_, columnGap))}
+     */
+    public Flex columnGap(Gap columnGap) {
+        return columnGap(breakpoints(default_, columnGap));
+    }
+
     /** Gap between columns at various breakpoints. This will override spacers for the main axis. */
     public Flex columnGap(Breakpoints<Gap> columnGap) {
         return css(columnGap.modifiers(gap -> "column-" + gap.value()));
+    }
+
+    /**
+     * Same as {@code direction(breakpoints(default_, direction))}
+     */
+    public Flex direction(Direction direction) {
+        return direction(breakpoints(default_, direction));
     }
 
     /** Value to add for flex-direction property at various breakpoints */
@@ -96,14 +140,35 @@ public class Flex extends BaseLayout<HTMLElement, Flex> {
         return css(direction.modifiers());
     }
 
+    /**
+     * Same as {@code display(breakpoints(default_, display))}
+     */
+    public Flex display(Display display) {
+        return display(breakpoints(default_, display));
+    }
+
     /** Value to set to display property at various breakpoints */
     public Flex display(Breakpoints<Display> display) {
         return css(display.modifiers());
     }
 
+    /**
+     * Same as {@code flex(breakpoints(default_, flexShorthand))}
+     */
+    public Flex flex(FlexShorthand flexShorthand) {
+        return flex(breakpoints(default_, flexShorthand));
+    }
+
     /** Value to add for flex property at various breakpoints */
     public Flex flex(Breakpoints<FlexShorthand> flexShorthand) {
         return css(flexShorthand.modifiers());
+    }
+
+    /**
+     * Same as {@code flexWrap(breakpoints(default_, flexWrap))}
+     */
+    public Flex flexWrap(FlexWrap flexWrap) {
+        return flexWrap(breakpoints(default_, flexWrap));
     }
 
     /** Value to set for flex-wrap property at various breakpoints */
@@ -119,6 +184,13 @@ public class Flex extends BaseLayout<HTMLElement, Flex> {
         return css(modifiers);
     }
 
+    /**
+     * Same as {@code gap(breakpoints(default_, gap))}
+     */
+    public Flex gap(Gap gap) {
+        return gap(breakpoints(default_, gap));
+    }
+
     /** Gap between items at various breakpoints. This will override spacers for the main axis. */
     public Flex gap(Breakpoints<Gap> gap) {
         return css(gap.modifiers());
@@ -132,13 +204,34 @@ public class Flex extends BaseLayout<HTMLElement, Flex> {
         return css(modifiers);
     }
 
+    /**
+     * Same as {@code justifyContent(breakpoints(default_, justifyContent))}
+     */
+    public Flex justifyContent(JustifyContent justifyContent) {
+        return justifyContent(breakpoints(default_, justifyContent));
+    }
+
     /** Value to add for justify-content property at various breakpoints */
     public Flex justifyContent(Breakpoints<JustifyContent> justifyContent) {
         return css(justifyContent.modifiers());
     }
 
+    /**
+     * Same as {@code order(breakpoints(default_, order))}
+     */
+    public Flex order(String order) {
+        return order(breakpoints(default_, order));
+    }
+
     public Flex order(Breakpoints<String> order) {
         return componentVar(layout(flex), item, Order).applyTo(this).set(order);
+    }
+
+    /**
+     * Same as {@code rowGap(breakpoints(default_, rowGap))}
+     */
+    public Flex rowGap(Gap rowGap) {
+        return rowGap(breakpoints(default_, rowGap));
     }
 
     /** Gap between rows at various breakpoints. This will override spacers for the main axis. */
@@ -154,9 +247,23 @@ public class Flex extends BaseLayout<HTMLElement, Flex> {
         return css(modifiers);
     }
 
+    /**
+     * Same as {@code spaceItems(breakpoints(default_, spaceItems))}
+     */
+    public Flex spaceItems(SpaceItems spaceItems) {
+        return spaceItems(breakpoints(default_, spaceItems));
+    }
+
     /** Space items at various breakpoints */
     public Flex spaceItems(Breakpoints<SpaceItems> spaceItems) {
         return css(spaceItems.modifiers());
+    }
+
+    /**
+     * Same as {@code spacer(breakpoints(default_, spacer))}
+     */
+    public Flex spacer(Spacer spacer) {
+        return spacer(breakpoints(default_, spacer));
     }
 
     /** Spacers at various breakpoints */