+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add shortcut methods for default breakpoints
+- Add `List.addItems(Iterable<T> items, Function<T, ListItem> display)` to add multiple list items
+
 ## [0.2.5] - 2024-05-21
 
 ### Added